+++ CHANGES.md
@@ -6,6 +6,7 @@ videojs-record changelog
 
 - Fix ES export syntax for `Record` (#330 by @extronics)
 - Bump required version for videojs-wavesurfer to 2.7.0 for `playToggle` fixes
+- Fix hiding `playToggle` control
 
 
 3.4.0 - 2019/02/04